+++ lib/generators/anyway/install/install_generator.rb
@@ -7,6 +7,8 @@ module Anyway
     class InstallGenerator < ::Rails::Generators::Base
       source_root File.expand_path("templates", __dir__)
 
+      class_option :configs_path, type: :string
+
       def copy_application_config
         template "application_config.rb", File.join(static_config_root, "application_config.rb")
       end
@@ -21,7 +23,8 @@ module Anyway
         inject_into_file "config/application.rb", after: %r{< Rails::Application\n} do
           <<-RUBY
     # Configure the path for configuration classes that should be used before initialization
-    # config.anyway_config.autoload_static_config_path = "#{static_config_root}"
+    # NOTE: path should be relative to the project root (Rails.root)
+    #{default_configs_path? ? "# " : ""}config.anyway_config.autoload_static_config_path = "#{static_config_root}"
           RUBY
         end
       end
@@ -29,7 +32,11 @@ module Anyway
       private
 
       def static_config_root
-        Anyway::Settings.autoload_static_config_path || "config/configs"
+        options[:configs_path] || Anyway::Settings.autoload_static_config_path || Anyway::DEFAULT_CONFIGS_PATH
+      end
+
+      def default_configs_path?
+        static_config_root == (Anyway::Settings.autoload_static_config_path || Anyway::DEFAULT_CONFIGS_PATH)
       end
     end
   end