+++ spec/dummy/app/configs/heroku_config.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class HerokuConfig < ApplicationConfig
+  attr_config :app_id, :app_name, :dyno_id, :release_version, :slug_commit
+end