+++ spec/dummy/config/application.rb
@@ -20,6 +20,9 @@ module Dummy
     config.autoloader = :zeitwerk if defined?(::Zeitwerk)
 
     config.anyway_config.use_local_files = false
+    if ENV["USE_APP_CONFIGS"] == "1"
+      config.anyway_config.autoload_static_config_path = "app/configs"
+    end
 
     config.heroku = HerokuConfig.instance